캐비넷 추가 기능 구현, 캐비넷 기능 수정, 예외처리 수정.

+++ src/main/java/com/minji/idusbackend/product/model/GetProductRes.java
@@ -4,16 +4,18 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigInteger;
+
 @Getter
 @Setter
 @AllArgsConstructor
 public class GetProductRes {
-    private int idx;
+    private BigInteger idx;
     private String name;
-    private int brandIdx;
-    private int categoryIdx;
-    private int price;
-    private int salePrice;
+    private BigInteger brandIdx;
+    private BigInteger categoryIdx;
+    private Integer price;
+    private Integer salePrice;
     private String deliveryType;
     private String isTodayDeal;
     private String filename;
